# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para memorias de cálculo de instalaciones eléctricas según normativa NOM (México)",
        "title": "Garfex Calculadora de Filtros API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/calculos/charola/espaciado": {
            "post": {
                "description": "Calcula el espaciado requerido entre conductores en charola tipo escalera",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Canalización"
                ],
                "summary": "Calcular espaciado en charola",
                "parameters": [
                    {
                        "description": "Datos de hilos y diámetro de conductores",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CharolaEspaciadoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espaciado calculado",
                        "schema": {
                            "$ref": "#/definitions/http.CharolaEspaciadoResponse"
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "schema": {
                            "$ref": "#/definitions/http.CharolaEspaciadoResponseError"
                        }
                    },
                    "422": {
                        "description": "No se pudo calcular el tamaño de charola",
                        "schema": {
                            "$ref": "#/definitions/http.CharolaEspaciadoResponseError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/http.CharolaEspaciadoResponseError"
                        }
                    }
                }
            }
        },
        "/calculos/charola/triangular": {
            "post": {
                "description": "Calcula el ancho de charola en configuración triangular (cables en posición Herculana)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Canalización"
                ],
                "summary": "Calcular configuración triangular",
                "parameters": [
                    {
                        "description": "Datos de hilos y diámetro de conductores",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CharolaTriangularRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuración triangular calculada",
                        "schema": {
                            "$ref": "#/definitions/http.CharolaTriangularResponse"
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "schema": {
                            "$ref": "#/definitions/http.CharolaTriangularResponseError"
                        }
                    },
                    "422": {
                        "description": "No se pudo calcular el tamaño de charola",
                        "schema": {
                            "$ref": "#/definitions/http.CharolaTriangularResponseError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/http.CharolaTriangularResponseError"
                        }
                    }
                }
            }
        },
        "/calculos/conductor-alimentacion": {
            "post": {
                "description": "Selecciona el conductor de alimentación (fase) según la corriente ajustada y tipo de canalización",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conductor"
                ],
                "summary": "Seleccionar conductor de alimentación",
                "parameters": [
                    {
                        "description": "Datos de corriente y canalización",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ConductorAlimentacionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conductor seleccionado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/http.ConductorAlimentacionResponse"
                        }
                    },
                    "400": {
                        "description": "Error de validación o datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/http.ConductorAlimentacionResponseError"
                        }
                    },
                    "422": {
                        "description": "No se encontró conductor adecuado",
                        "schema": {
                            "$ref": "#/definitions/http.ConductorAlimentacionResponseError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/http.ConductorAlimentacionResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CharolaEspaciadoOutput": {
            "type": "object",
            "properties": {
                "ancho_comercial_mm": {
                    "type": "number"
                },
                "ancho_control_mm": {
                    "type": "number"
                },
                "ancho_fuerza_mm": {
                    "type": "number"
                },
                "ancho_requerido_mm": {
                    "type": "number"
                },
                "ancho_tierra_mm": {
                    "type": "number"
                },
                "diametro_control_mm": {
                    "type": "number"
                },
                "diametro_fase_mm": {
                    "description": "Valores intermedios del desarrollo — para mostrar en memoria de cálculo",
                    "type": "number"
                },
                "diametro_tierra_mm": {
                    "type": "number"
                },
                "espacio_control_mm": {
                    "type": "number"
                },
                "espacio_fuerza_mm": {
                    "type": "number"
                },
                "num_hilos_total": {
                    "type": "integer"
                },
                "tamano": {
                    "type": "string"
                },
                "tamano_pulgadas": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                }
            }
        },
        "dto.CharolaTriangularOutput": {
            "type": "object",
            "properties": {
                "ancho_comercial_mm": {
                    "type": "number"
                },
                "ancho_control_mm": {
                    "type": "number"
                },
                "ancho_potencia_mm": {
                    "type": "number"
                },
                "ancho_requerido_mm": {
                    "type": "number"
                },
                "ancho_tierra_mm": {
                    "type": "number"
                },
                "diametro_control_mm": {
                    "type": "number"
                },
                "diametro_fase_mm": {
                    "description": "Valores intermedios del desarrollo — para mostrar en memoria de cálculo",
                    "type": "number"
                },
                "diametro_tierra_mm": {
                    "type": "number"
                },
                "espacio_control_mm": {
                    "type": "number"
                },
                "espacio_fuerza_mm": {
                    "type": "number"
                },
                "factor_triangular": {
                    "type": "number"
                },
                "tamano": {
                    "type": "string"
                },
                "tamano_pulgadas": {
                    "type": "string"
                },
                "tipo": {
                    "type": "string"
                }
            }
        },
        "dto.ConductorAlimentacionOutput": {
            "type": "object",
            "properties": {
                "calibre": {
                    "description": "Calibre es el calibre del conductor seleccionado (ej: \"4 AWG\", \"250 MCM\").",
                    "type": "string"
                },
                "capacidad_nominal": {
                    "description": "CapacidadNominal es la ampacidad del conductor segun la tabla.",
                    "type": "number"
                },
                "material": {
                    "description": "Material es el material del conductor (\"Cu\" o \"Al\").",
                    "type": "string"
                },
                "seccion_mm2": {
                    "description": "SeccionMM2 es la seccion transversal en mm2.",
                    "type": "number"
                },
                "tabla_usada": {
                    "description": "TablaUsada es el nombre descriptivo de la tabla NOM utilizada.",
                    "type": "string"
                },
                "tipo_aislamiento": {
                    "description": "TipoAislamiento es el tipo de aislamiento (ej: \"THHN/THHW\").",
                    "type": "string"
                }
            }
        },
...
        "http.CharolaEspaciadoRequest": {
            "type": "object",
            "required": [
                "diametro_fase_mm",
                "diametro_tierra_mm",
                "hilos_por_fase",
                "sistema_electrico"
            ],
            "properties": {
                "diametro_control_mm": {
                    "type": "number"
                },
                "diametro_fase_mm": {
                    "type": "number"
                },
                "diametro_tierra_mm": {
                    "type": "number"
                },
                "hilos_por_fase": {
                    "type": "integer"
                },
                "sistema_electrico": {
                    "type": "string"
                }
            }
        },
        "http.CharolaEspaciadoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CharolaEspaciadoOutput"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.CharolaEspaciadoResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.CharolaTriangularRequest": {
            "type": "object",
            "required": [
                "diametro_fase_mm",
                "diametro_tierra_mm",
                "hilos_por_fase"
            ],
            "properties": {
                "diametro_control_mm": {
                    "type": "number"
                },
                "diametro_fase_mm": {
                    "type": "number"
                },
                "diametro_tierra_mm": {
                    "type": "number"
                },
                "hilos_por_fase": {
                    "type": "integer"
                }
            }
        },
        "http.CharolaTriangularResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CharolaTriangularOutput"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.CharolaTriangularResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.ConductorAlimentacionRequest": {
            "type": "object",
            "required": [
                "corriente_ajustada",
                "tipo_canalizacion"
            ],
            "properties": {
                "corriente_ajustada": {
                    "type": "number"
                },
                "hilos_por_fase": {
                    "type": "integer"
                },
                "material": {
                    "type": "string"
                },
                "temperatura": {
                    "type": "integer"
                },
                "tipo_canalizacion": {
                    "type": "string"
                }
            }
        },
        "http.ConductorAlimentacionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ConductorAlimentacionOutput"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.ConductorAlimentacionResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}